+++ src/config/endpoints.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * @link https://github.com/panlatent/craft-aliyun
+ * @copyright Copyright (c) 2018 panlatent@gmail.com
+ */
+
+return [
+    ['name' => '华东 1（杭州）', 'public' => 'oss-cn-hangzhou.aliyuncs.com', 'internal' => 'oss-cn-hangzhou-internal.aliyuncs.com'],
+    ['name' => '华东 2（上海）', 'public' => 'oss-cn-shanghai.aliyuncs.com', 'internal' => 'oss-cn-shanghai-internal.aliyuncs.com'],
+    ['name' => '华北 1（青岛）', 'public' => 'oss-cn-qingdao.aliyuncs.com', 'internal' => 'oss-cn-qingdao-internal.aliyuncs.com'],
+    ['name' => '华北 2（北京）', 'public' => 'oss-cn-beijing.aliyuncs.com', 'internal' => 'oss-cn-beijing-internal.aliyuncs.com'],
+    ['name' => '华北 3（张家口）', 'public' => 'oss-cn-zhangjiakou.aliyuncs.com', 'internal' => 'oss-cn-zhangjiakou-internal.aliyuncs.com'],
+    ['name' => '华北 5（呼和浩特）', 'public' => 'oss-cn-huhehaote.aliyuncs.com', 'internal' => 'oss-cn-huhehaote-internal.aliyuncs.com'],
+    ['name' => '华南 1（深圳）', 'public' => 'oss-cn-shenzhen.aliyuncs.com', 'internal' => 'oss-cn-shenzhen-internal.aliyuncs.com'],
+    ['name' => '香港', 'public' => 'oss-cn-hongkong.aliyuncs.com', 'internal' => 'oss-cn-hongkong-internal.aliyuncs.com'],
+    ['name' => '美国西部 1（硅谷）', 'public' => 'oss-us-west-1.aliyuncs.com', 'internal' => 'oss-us-west-1-internal.aliyuncs.com'],
+    ['name' => '美国东部 1（弗吉尼亚）', 'public' => 'oss-us-east-1.aliyuncs.com', 'internal' => 'oss-us-east-1-internal.aliyuncs.com'],
+    ['name' => '亚太东南 1（新加坡）', 'public' => 'oss-ap-southeast-1.aliyuncs.com', 'internal' => 'oss-ap-southeast-1-internal.aliyuncs.com'],
+    ['name' => '亚太东南 2（悉尼）', 'public' => 'oss-ap-southeast-2.aliyuncs.com', 'internal' => 'oss-ap-southeast-2-internal.aliyuncs.com'],
+    ['name' => '亚太东南 3（吉隆坡）', 'public' => 'oss-ap-southeast-3.aliyuncs.com', 'internal' => 'oss-ap-southeast-3-internal.aliyuncs.com'],
+    ['name' => '亚太东南 5（雅加达）', 'public' => 'oss-ap-southeast-5.aliyuncs.com', 'internal' => 'oss-ap-southeast-5-internal.aliyuncs.com'],
+    ['name' => '亚太东北 1（日本）', 'public' => 'oss-ap-northeast-1.aliyuncs.com', 'internal' => 'oss-ap-northeast-1-internal.aliyuncs.com'],
+    ['name' => '亚太南部 1（孟买）', 'public' => 'oss-ap-south-1.aliyuncs.com', 'internal' => 'oss-ap-south-1-internal.aliyuncs.com'],
+    ['name' => '欧洲中部 1（法兰克福）', 'public' => 'oss-eu-central-1.aliyuncs.com', 'internal' => 'oss-eu-central-1-internal.aliyuncs.com'],
+    ['name' => '英国（伦敦）', 'public' => 'oss-eu-west-1.aliyuncs.com', 'internal' => 'oss-eu-west-1-internal.aliyuncs.com'],
+    ['name' => '中东东部 1（迪拜）', 'public' => 'oss-me-east-1.aliyuncs.com', 'internal' => 'oss-me-east-1-internal.aliyuncs.com'],
+];