+++ src/translations/zh/aliyun.php
@@ -1,8 +1,6 @@
 <?php
 /**
- * Aliyun plugin for Craft 3
- *
- * @link      https://panlatent.com/
+ * @link https://github.com/panlatent/craft-aliyun
  * @copyright Copyright (c) 2018 panlatent@gmail.com
  */
 
@@ -34,4 +32,9 @@ return [
     'Storage space name' => '该空间的 Bucket 名称',
     'This bucket is public space?' => '这个是一个公开访问的存储空间？',
     'The volume file will be saved to the subdirectory of this Bucket' => '这个卷文件将保存在该路径下, 需要与公开卷 URL 路径对应',
+    'Use environment variables storage {Access Key} and {Secret Key}' => '使用环境变量存储 {Access Key} 和 {Secret Key}',
+    'Use Global Settings' => '使用全局设置',
+    'Use global plugin settings. This requires you to set the plugin settings' => '使用全局的插件设置，这需要你配置插件设置',
+    'Endpoint List' => '节点列表',
+    '(Internal)' => '（ 内网 ）',
 ];